+++ database/deleteUser.php
@@ -1,7 +1,7 @@
 <?php
     include 'connection.php';
 
-    if(!empty($_GET['id'])){
+    if(!empty($_GET['id']) && is_numeric($_GET['id'])){
         $stmt = $db->prepare('DELETE FROM USER WHERE user_id = :user_id');
         $stmt->bindParam(':user_id', $_GET['id']);
         $stmt->execute();
